Custom JsonBooleanConverter to handle strings
The UI is sending in "true"/"false"/"null" as strings and the "null" is
throwing off the JSON conversion.  This allows the calls to be made
using things that look like true (true/yes/y/1) or false (false/no/n/0)
with the fallback to NULL for anything it doesn't understand.